refactor : separate Basenews logic and Dailynews logic
